app/python_scraper/requirements.txt
playwright
langdetect
vaderSentiment
spacy
kafka-python
aiokafka
aiohttp
python-dotenv